Split up the fragment sink
